Fleshed out movieSelection.js
My code now assumes that 'movieSelection/home' will be the landing page for decision rooms.

This is slightly different from before, but there needs to be a page/handlebars layout to send the user back to after the decision process is over, which should likely be the decision room home (will display final movie).